// frontend/types/api/shipments/models/PackageStatus.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum PackageStatus {
    PENDING = 'PENDING',
    IN_TRANSIT = 'IN_TRANSIT',
    OUT_FOR_DELIVERY = 'OUT_FOR_DELIVERY',
    DELIVERED = 'DELIVERED',
    EXCEPTION = 'EXCEPTION',
    DELAYED = 'DELAYED',
    CANCELLED = 'CANCELLED',
    RETURNED = 'RETURNED',
}